Add files via upload
Provide all details of your extension’s features and functionality. Include any updates.

AJAX Advanced Search extension is a Magento specific tool that modifies the advanced search feature. It allows a user to view the result of a search on the same page, instead of displaying the result of a search on the next page.

This extension also allows admin, to enable/disable the configurable attributes option (Product variant) of the products in the search, from admin panel. In case it is enabled; the user can add product variant from the search result, and if it is disabled; different attributes of the product are no longer displayed in search results.

The AJAX Advanced Search extension can also be disabled from the admin panel to revert to default functionality of advanced search. Read User Guide to learn how to add "AJAX Advanced Search" extension to your Magento Store, and how to enable/disable different features.  This extension is free for all Magento 2 users.

 

Features:
Add AJAX feature in the advanced search
AJAX feature can be enabled/disabled
 If enabled; AJAX Advanced Search result will display on the same page without page reload
If disabled; AJAX Advanced Search results will display on a new page as in default functionality
Configurable options can be enabled/disabled for Search results
 If configurable option is enabled; the user can add a variant of the product to the cart from AJAX Advanced Search results
If configurable option is disabled; Variant of a product is not visible for AJAX Advanced Search results
The search form becomes minimized and is replaced by the Search Again button, while the AJAX Advanced Search results are displayed
The search form can be opened again by clicking on the Search Again button